fix bug total price cart slice
